Upgrade to Unity 5.5.1p2, fix fog of war, tweak stuff

+++ Assets/Scripts/Health/Health.cs
@@ -2,6 +2,6 @@
 
 public class Health : MonoBehaviour
 {
-    public int maxHealth = 1000;
-    public int currentHealth = 1000;
+    public float maxHealth = 1000;
+    public float currentHealth = 1000;
 }
